+++ lib/alien_signals.dart
@@ -1,2 +1,3 @@
-export 'src/preset.dart' show startBatch, endBatch, trigger;
+export 'src/preset.dart'
+    show EffectCallback, EffectCleanup, startBatch, endBatch, trigger;
 export 'src/surface.dart';